modles_usage/src/requirements.txt
fastapi
uvicorn
ollama
openai
google.generativeai
llama-cpp-python
transformers
accelerate
sentencepiece
